message

+++ WebAPI/WebAppApi/src/main/java/com/example/WebAppApi/Repository/businessRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.WebAppApi.Model.Business;
 
-public interface BusinessRepository extends JpaRepository<Business, Integer> {
+public interface BusinessRepository extends JpaRepository<Business, Long> {
     
 }
